fix: reconnecting to the service-provider
